+++ src/components/containers/MidColumn.tsx
@@ -1,9 +1,9 @@
 import styled from "styled-components";
 import {} from "../particles/Buttons";
 import StoryContainer from "./MidColumn/Story/StoryContainer";
-import PostContainer from "./MidColumn/Post/PostMasterContainer";
 import { useEffect, useState } from "react";
 import PostsData from "../../../public/FakeAPI/Post/PostType";
+import ComPost from "./MidColumn/Post/Post";
 const MidColumn = styled.div`
   width: 50%;
   border: 1px dashed blue;
@@ -19,10 +19,10 @@ const PostsCollection = () => {
   }, [setPagePosts]);
 
   return PagePosts.Posts.map((post) => {
-    return <PostContainer post={post} key={post.ID}></PostContainer>;
+    return <ComPost post={post} key={post.ID}></ComPost>;
   });
 };
-
+//
 export default function MidColumnComponent() {
   return (
     <MidColumn className="MidColumn">
